requirements.txt
openai>=1.0.0
mcp[cli]>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
tqdm>=4.65.0
